Add Select Videos to album import menu and delete unused ImporterView
- Add "Select Videos" option to the album toolbar import menu
- Add video import support to AlbumView (importSelectedVideos,
  importLoadedFiles with video detection, importFiles with video)
- Add videos picker state and .photosPicker modifier for videos
- Delete unused ImporterView.swift
